requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.2
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
imageio==2.6.1
isort==4.3.21
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
matplotlib==3.1.1
mccabe==0.6.1
mock==3.0.5
networkx==2.4
numpy==1.16.4
opt-einsum==3.1.0
Pillow==6.2.1
progressbar==2.5
protobuf==3.10.0
pydot==1.4.1
pylint==2.4.3
pyparsing==2.4.2
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.1.2
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
